embedded/RaspberryPi/requirements.txt
absl-py==2.3.1
aioice==0.10.1
aiortc==1.14.0
attrs==25.4.0
av==16.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cv2_enumerate_cameras==1.3.0
cycler==0.12.1
dnspython==2.8.0
dotenv==0.9.9
flatbuffers==25.9.23
fonttools==4.60.1
google-crc32c==1.7.1
idna==3.11
ifaddr==0.2.0
kiwisolver==1.4.9
matplotlib==3.10.7
mediapipe==0.10.9
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
protobuf==3.20.3
pvporcupine==3.0.5
pvrecorder==1.2.7
PyAudio==0.2.14
pycparser==2.23
pydub==0.25.1
pyee==13.0.0
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyopenvidu==0.2.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sounddevice==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
webrtcvad==2.0.10
websockets==15.0.1
